Various fixes: use fork of go-plugins-helpers, Functionbeat, templates (#16191)
## What does this PR do?

From now on our for of `github.com/docker/go-plugins-helpers` is used until that lib migrates to modules.

Furthermore, vendoring dependencies for GCP functions is fixed. In modules mode, the output of `go list -deps` does not contain the prefix `github.com/elastic/beats/vendor`. As it does not look for packages under vendor.

As Beat generators were moved to the folder `_templates`, the license header check needed adjustment to exclude the templates.

## Why is it important?

Dockerlogbeat could not be built because of the lib is not a module.

`make update` runs successfully under `x-pack/functionbeat`.